Fix font rendering, make compose DSL safe

+++ src/main/java/io/karma/pda/api/client/render/graphics/FontAtlas.java
@@ -33,15 +33,15 @@ public interface FontAtlas extends Disposable {
 
     int getSpriteBorder();
 
-    int getMaxGlyphWidth();
+    float getMaxGlyphWidth();
 
-    int getMaxGlyphHeight();
+    float getMaxGlyphHeight();
 
-    int getMaxGlyphBearingX();
+    float getMaxGlyphBearingX();
 
-    int getMaxGlyphBearingY();
+    float getMaxGlyphBearingY();
 
-    int getLineHeight();
+    float getLineHeight();
 
     int getTextureId();
