+++ src/main/java/io/karma/pda/api/client/render/graphics/GlyphMetrics.java
@@ -13,19 +13,19 @@ import net.minecraftforge.api.distmarker.OnlyIn;
  */
 @OnlyIn(Dist.CLIENT)
 public interface GlyphMetrics {
-    int getWidth();
+    float getWidth();
 
-    int getHeight();
+    float getHeight();
 
-    int getAscent();
+    float getAscent();
 
-    int getDescent();
+    float getDescent();
 
-    int getAdvanceX();
+    float getAdvanceX();
 
-    int getAdvanceY();
+    float getAdvanceY();
 
-    int getBearingX();
+    float getBearingX();
 
-    int getBearingY();
+    float getBearingY();
 }
